"Updated auth controller to return 401 status for duplicate user data, fixed deleteUser function, added confirmation prompt for expelStudent, and made minor changes to application form HTML and JavaScript."

+++ public/admin/allstudent.js
@@ -33,15 +33,17 @@ function expelStudent(student) {
  
      console.log("Delete URL:", deleteUrl); // Log the URL
  
-     axios.delete(deleteUrl)
-         .then((response) => {
-             alert('Student expelled successfully!');
-             location.reload();
-         })
-         .catch((error) => {
-                console.error("Error expelling student:", error);
-          }) 
- }
+     if (confirm(`Are you sure you want to Expel ${student.firstName} ${student.lastName}`)) {
+          axios.delete(deleteUrl)
+          .then((response) => {
+               alert('Student expelled successfully!');
+               location.reload();
+          })
+          .catch((error) => {
+                    console.error("Error expelling student:", error);
+               }) 
+          }
+     }
  
  // Adjust the showStudentDetails function to remove the old event listener
  function showStudentDetails(student) {
